Migrate lint-self-disarmable-gate to engine corpus preparation
Extract AST scanner into self_disarmable_gate_detector.py with corpus-based
prepare_corpus/detect; rewrite wrapper to check for Piece 1's prepare_corpus
contract and fail closed when absent; update tests to use SourceFile fixtures.
